eval-client/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "eval-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Apoorva Sareen",email = "apoorva@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11.9,<3.12"
dependencies = [
    "langfuse (>=3.10.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "psutil (>=5.9.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
